Move DefaultSecureTransportParameters to separate file and add javadoc
